% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_edm.R
\name{easy_edm}
\alias{easy_edm}
\title{easy_edm}
\usage{
easy_edm(
  cause,
  effect,
  time = NULL,
  data = NULL,
  direction = "oneway",
  verbosity = 0
)
}
\arguments{
\item{cause}{The first time series in the causal analysis}

\item{effect}{The second time series in the causal analysis}

\item{time}{If time is not uniformly sampled, then it must be supplied here.}

\item{data}{If a dataframe is supplied here, then cause, effect & time must
be strings containing the column names of the relevant time series.}

\item{direction}{A string specifying whether we are checking a one
directional causal effect or whether to test the reverse direction at the
same time.}

\item{verbosity}{The level of detail in the output.}
}
\value{
list
}
\description{
easy_edm
}
