\name{fastEDM-package}
\alias{fastEDM-package}
\alias{fastEDM}
\docType{package}
\title{
\packageTitle{fastEDM}
}
\description{
\packageDescription{fastEDM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fastEDM}
\packageIndices{fastEDM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fastEDM}

Maintainer: \packageMaintainer{fastEDM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
