% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_map.R
\name{logistic_map}
\alias{logistic_map}
\title{logistic_map}
\usage{
logistic_map(obs)
}
\arguments{
\item{obs}{The number of observations to generate.}
}
\value{
A list containing the synthetic x & y time series.
}
\description{
This function generates a synthetic time series called the logistic map.
It can be used to perform basic tests of EDM causality.
}
