% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_edm.R
\name{easy_edm}
\alias{easy_edm}
\title{easy_edm}
\usage{
easy_edm(
  cause,
  effect,
  time = NULL,
  data = NULL,
  direction = "oneway",
  verbosity = 1,
  showProgressBar = NULL,
  normalize = TRUE
)
}
\arguments{
\item{cause}{The causal time series (as a string or a vector).}

\item{effect}{The effect time series (as a string or a vector).}

\item{time}{For non-regularly sampled time series, the sampling times
must be supplied here (as a string or a vector).}

\item{data}{If a dataframe is supplied here, then the cause, effect & time
arguments must be the column names of the relevant time series as strings.}

\item{direction}{A string specifying whether we are checking a one
directional causal effect or whether to test the reverse direction at the
same time (work in progress!).}

\item{verbosity}{The level of detail in the output.}

\item{showProgressBar}{Whether or not to print out a progress bar during the computations.}

\item{normalize}{Whether to normalize the inputs before starting EDM.}
}
\value{
A Boolean indicating that evidence of causation was found.
}
\description{
easy_edm
}
\examples{
library(fastEDM)

chicago <- read.csv(url("https://github.com/EDM-Developers/fastEDM-r/raw/main/vignettes/chicago.csv"))

crimeCCMCausesTemp <- easy_edm("Crime", "Temperature", data=chicago, verbosity=0)
tempCCMCausesCrime <- easy_edm("Temperature", "Crime", data=chicago, verbosity=0)
}
