% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastEDM.R
\docType{package}
\name{fastEDM}
\alias{fastEDM}
\title{fastEDM}
\description{
The fastEDM package implements a series of tools that can be
used for empirical dynamic modeling. The core algorithm is written in C++
to achieve a reasonable execution speed.
}
\examples{
library(fastEDM)

chicago <- read.csv(url("https://github.com/EDM-Developers/fastEDM/raw/master/vignettes/chicago.csv"))

crimeCCMCausesTemp <- easy_edm("Crime", "Temperature", data=chicago, verbosity=0)
tempCCMCausesCrime <- easy_edm("Temperature", "Crime", data=chicago, verbosity=0)
}
\references{
Jinjing Li, Michael J. Zyphur, George Sugihara, Patrick J. Laub (2021), Beyond Linearity, Stability, and Equilibrium: The edm Package for Empirical Dynamic Modeling and Convergent Cross Mapping in Stata, Stata Journal, 21(1), pp. 220-258
}
\seealso{
{
  \url{https://edm-developers.github.io/fastEDM/} and \url{https://edm-developers.github.io/EDM/}
}
}
\author{
Patrick Laub <patrick.laub@gmail.com>
}
